Moves BP recording form back to components folder
