+++ components/BloodPressureRecordingForm.tsx
@@ -2,15 +2,15 @@ import { useNavigation, NavigationProp, Route } from '@react-navigation/native';
 import { useContext, useEffect, useRef, useState } from 'react';
 import { StyleSheet, Text, TextInput, TouchableOpacity, View } from 'react-native';
 import DateTimePickerModal from "react-native-modal-datetime-picker";
-import { AppStackParamList } from '../../App';
-import { BloodPressureDispatchAction, BloodPressureDispatchActionType } from '../../data/BloodPressureDispatchAction';
-import { BloodPressureRecordingDispatchContext } from '../../data/BloodPressureRecordingProvider';
-import { selectBloodPressureToEdit } from '../../hooks/selectBloodPressureToEdit';
-import { useErrorString } from '../../hooks/useErrorString';
-import { useIncrementingDateTime } from '../../hooks/useIncrementingDate';
-import { useNumberState } from '../../hooks/useNumberState';
-import { BloodPressureRecording } from '../../models/BloodPressureRecording';
-import { formatDateAsYYYYMMDD, formatTimeFromDate } from '../../models/conversions';
+import { AppStackParamList } from '../App';
+import { BloodPressureDispatchAction, BloodPressureDispatchActionType } from '../data/BloodPressureDispatchAction';
+import { BloodPressureRecordingDispatchContext } from '../data/BloodPressureRecordingProvider';
+import { selectBloodPressureToEdit } from '../hooks/selectBloodPressureToEdit';
+import { useErrorString } from '../hooks/useErrorString';
+import { useIncrementingDateTime } from '../hooks/useIncrementingDate';
+import { useNumberState } from '../hooks/useNumberState';
+import { BloodPressureRecording } from '../models/BloodPressureRecording';
+import { formatDateAsYYYYMMDD, formatTimeFromDate } from '../models/conversions';
 
 export interface IBloodPressureRecordingFormRouteParams {
   bloodPressureRecordingIdToEdit?: string
